Fix MASSIVE Security Bugs, Update Tests and More
- Fixed MASSIVE SECURITY bug where ANY Authenticated User with the
required Authorized Role and Permissions could perform CRUD operations
on another Organization's: Users, Projects, Issues, Comments,
and Files, without having any connection to the Requested Organization

- Fixed Security bug where ANY Authenticated User was allowed to VIEW
and UPDATE ANY User's 'Account Profile' if they had the correct User
ID number. Now Users can only access their own 'Account Profile' page

- Also fixed the same bugs for the 'Account Security' and
'Account Notifications' pages

- Fixed bug where if two or more Projects had the same Identifier, it
would crash an associated webpage because the Database Query would not
return a unique result. this was fixed by changing the Query to not
only the Project Identifier but to an Organization as well. To make
this work, I also changed the Project Creation to always have UNIQUE
Project Identifiers and Access Keys within an Organization

- Fixed bug where if two or more Issues had the same Project Identifier
would crash the 'View Issue details' page because sometimes it would
not return a UNIQUE result. this was resolved by changing the Query to
find the Issue by Issue Key AND Organization. also, adding a column to
the IssueKeySequence DB table to hold the Project's ID created UNIQUE
Issue Queries

- Updated some Test Cases to accommodate for the Project and Issue
Query changes

- Changed the way a List of an Organization's Users by Role was being
queried

- Added permissions of deleting Files only to Project Managers, Admins,
and Developers

- Removed access to CREATE and/or UPDATE an 'Issue' from Role: 'Guest'

- Added more popup alerts for CRUD operations on Projects

- Also cleaned up some code and added/modified documentation for
readability and uniformity

+++ src/main/java/com/sephrael/issuetrackingsystem/controller/CommentController.java
@@ -2,9 +2,11 @@ package com.sephrael.issuetrackingsystem.controller;
 
 import com.sephrael.issuetrackingsystem.entity.Comment;
 import com.sephrael.issuetrackingsystem.entity.Issue;
+import com.sephrael.issuetrackingsystem.entity.Project;
 import com.sephrael.issuetrackingsystem.entity.User;
 import com.sephrael.issuetrackingsystem.repository.CommentRepository;
 import com.sephrael.issuetrackingsystem.repository.IssueRepository;
+import com.sephrael.issuetrackingsystem.repository.ProjectRepository;
 import com.sephrael.issuetrackingsystem.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -21,19 +23,28 @@ public class CommentController {
     private IssueRepository issueRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private ProjectRepository projectRepository;
 
     @PostMapping("/new")
     public String addComment(@PathVariable(value = "issueKey") String issueKey, Comment comment, Principal principal,
                              @PathVariable String identifier) {
+        User currentUser = userRepository.findByEmail(principal.getName());
+        Project currentProject = projectRepository.findByIdentifierAndOrganization(identifier, currentUser.getOrganization());
+
+        if(currentUser.getOrganization() == null)
+            return "/organization/select-organization";
+
+        // checks if the Current User is involved with the requested Project
+        if(!currentUser.getProjects().contains(currentProject))
+            return "/error/404";
+
         userRepository.findByEmail(principal.getName()).addToComment(comment);
-        issueRepository.findByIssueKey(issueKey).addToComment(comment);
+        issueRepository.findByIssueKeyAndOrganization(issueKey, currentUser.getOrganization()).addToComment(comment);
 
         comment.setIsEdited(false);
         commentRepository.save(comment);
 
-        if(userRepository.findByEmail(principal.getName()).getOrganization() == null) {
-            return "/organization/select-organization";
-        }
         return "redirect:/issues/{identifier}/view/" + issueKey;
     }
 
@@ -42,20 +53,28 @@ public class CommentController {
                               @RequestParam("isEdited") boolean isEdited, @RequestParam("message") String message,
                               @PathVariable(value = "issueKey") String issueKey,
                               @PathVariable String identifier, Principal principal) {
+        User currentUser = userRepository.findByEmail(principal.getName());
+        Project currentProject = projectRepository.findByIdentifierAndOrganization(identifier, currentUser.getOrganization());
+
+        if(currentUser.getOrganization() == null)
+            return "/organization/select-organization";
+
+        // checks if the Current User is involved with the requested Project
+        if(!currentUser.getProjects().contains(currentProject))
+            return "/error/404";
 
         Comment updatedComment = commentRepository.findCommentById(id);
 
         updatedComment.setUser(user);
         updatedComment.setIssue(issue);
+
         // sets Comment isEdited to true if comment has not yet been edited
         if(!isEdited) { updatedComment.setIsEdited(true); }
+
         updatedComment.setMessage(message);
 
         commentRepository.save(updatedComment);
 
-        if(userRepository.findByEmail(principal.getName()).getOrganization() == null) {
-            return "/organization/select-organization";
-        }
         return "redirect:/issues/{identifier}/view/" + issueKey;
     }
 
@@ -63,17 +82,27 @@ public class CommentController {
     public String deleteComment(@PathVariable(value = "issueKey") String issueKey,
                                 @PathVariable(value = "commentId") Long commentId,
                                 @PathVariable String identifier, Principal principal) {
+        User currentUser = userRepository.findByEmail(principal.getName());
+        Project currentProject = projectRepository.findByIdentifierAndOrganization(identifier, currentUser.getOrganization());
+
+        if(currentUser.getOrganization() == null)
+            return "/organization/select-organization";
+
+        // checks if the Current User is involved with the requested Project
+        if(!currentUser.getProjects().contains(currentProject))
+            return "/error/404";
+
+
         commentRepository.deleteById(commentId);
 
-        if(userRepository.findByEmail(principal.getName()).getOrganization() == null) {
-            return "/organization/select-organization";
-        }
         return "redirect:/issues/{identifier}/view/" + issueKey;
     }
 
     // this shows the json format of all the Comments of an Issue
     @GetMapping(path = "/all")
-    public @ResponseBody Iterable<Comment> getAllCommentsByIssueId(@PathVariable(value = "issueKey") String issueKey, @PathVariable(name = "identifier") String identifier) {
-        return issueRepository.findByIssueKey(issueKey).getComments();
+    public @ResponseBody Iterable<Comment> getAllCommentsByIssueId(@PathVariable(value = "issueKey") String issueKey,
+                                                                   @PathVariable(name = "identifier") String identifier, Principal principal) {
+        return issueRepository.findByIssueKeyAndOrganization(issueKey,
+                userRepository.findByEmail(principal.getName()).getOrganization()).getComments();
     }
 }
