+++ src/main/java/com/sephrael/issuetrackingsystem/controller/FileController.java
@@ -17,6 +17,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Controller
@@ -31,8 +32,18 @@ public class FileController {
     private UserRepository userRepository;
 
     @GetMapping("/{id}")
-    public ResponseEntity<byte[]> getFileById(@PathVariable long id) {
+    public ResponseEntity<byte[]> getFileById(@PathVariable long id, Principal principal) {
         File file = fileService.getFile(id);
+        User currentUser = userRepository.findByEmail(principal.getName());
+
+        // if the requested file is a Profile Picture AND if it matches the Current User's Organization, redirect to NOT FOUND page
+        if(file.isProfilePicture() && file.getUser().getOrganization() != currentUser.getOrganization())
+            return ResponseEntity.notFound().build();
+
+        // if the requested file is NOT a Profile Picture AND if the Current User is NOT involved with the
+        // File's associated 'Project', redirect to NOT FOUND page
+        if(!file.isProfilePicture() && !file.getIssue().getProject().getUsers().contains(currentUser))
+            return ResponseEntity.notFound().build();
 
         return ResponseEntity.ok()
                 .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.getName() + "\"")
@@ -40,8 +51,12 @@ public class FileController {
     }
 
     @GetMapping("/{userId}/{isProfilePicture}")
-    public ResponseEntity<byte[]> getProfilePicture(@PathVariable("userId") long userId,
+    public ResponseEntity<byte[]> getProfilePicture(@PathVariable("userId") long userId, Principal principal,
                                                     @PathVariable("isProfilePicture") boolean isProfilePicture) {
+        // if the requested File's 'Organization' does NOT match the Current User's 'Organization', redirect to NOT FOUND page
+        if(userRepository.findUserById(userId).getOrganization() != userRepository.findByEmail(principal.getName()).getOrganization())
+            return ResponseEntity.notFound().build();
+
         File file = fileRepository.findByUserAndIsProfilePicture(userRepository.findUserById(userId), isProfilePicture);
 
         return ResponseEntity.ok()
@@ -50,19 +65,32 @@ public class FileController {
     }
 
     @PostMapping("/upload/{userId}/{isProfilePicture}")
-    public String uploadFile(@PathVariable("userId") long userId, @RequestParam("file")MultipartFile file,
+    public String uploadFile(@PathVariable("userId") long userId, @RequestParam("file")MultipartFile file, Principal principal,
                                                       @PathVariable("isProfilePicture") boolean isProfilePicture) {
         User user = userRepository.findUserById(userId);
 
+        // if Current User does NOT match the requested User, redirect to 404 page
+        if(userRepository.findByEmail(principal.getName()).getId() != userId)
+            return "/error/404";
+
+        // if the 'User' uploaded an empty 'File' AND the 'User' already has a 'Profile Picture', DELETE the User's current
+        // 'Profile Picture' and SET the User's 'hasProfilePicture' to FALSE
         if(file.isEmpty() && user.hasProfilePicture()) {
             fileService.deleteFile(fileRepository.findByUserAndIsProfilePicture(user, isProfilePicture).getId());
 
             user.setHasProfilePicture(false);
             userRepository.save(user);
+
+        // if 'File' is empty and 'User' does NOT have a 'Profile Picture', redirect to 'Account Profile' page
         } else if(file.isEmpty() && !user.hasProfilePicture()) {
             return "redirect:/account/profile/" + userId;
-        } else
+
+        // if 'File' is NOT empty AND 'User' does NOT have a 'Profile Picture', upload and set the 'File' to the User's
+        // 'Profile Picture'
+        } else if(!file.isEmpty() && !user.hasProfilePicture()) {
             fileService.uploadFile(user, file, isProfilePicture);
+        } else
+            return "/error/400";
 
         return "redirect:/account/profile/" + userId;
     }
@@ -70,11 +98,17 @@ public class FileController {
     @RequestMapping("/{identifier}/{issueKey}/delete/{id}")
     public String deleteFile(@PathVariable("id") long id, @PathVariable("identifier") String identifier,
                              @PathVariable("issueKey") String issueKey, Principal principal) {
+        User currentUser = userRepository.findByEmail(principal.getName());
+
+        if(currentUser.getOrganization() == null)
+            return "/organization/select-organization";
+
+        // if Current User does NOT match the requested File's original Uploader, redirect to 404 page
+        if(!Objects.equals(currentUser.getId(), fileService.getFile(id).getUser().getId()))
+            return "/error/404";
+
         fileRepository.deleteById(id);
 
-        if(userRepository.findByEmail(principal.getName()).getOrganization() == null) {
-            return "/organization/select-organization";
-        }
         return String.format("redirect:/issues/%s/view/%s", identifier, issueKey);
     }
